[FLINK-4505][Cluster Management] Implement TaskManagerFactory to bring up TaskManager for different modes
Summary: above

Test Plan: NA

Reviewers: kete.yangkt

Subscribers: #blink

Differential Revision: http://phabricator.taobao.net/D5606
